+++ students/ScotchWSplenda/lesson09/assignment/charges_calcV2.py
@@ -1,6 +1,5 @@
 '''
 https://lerner.co.il/2019/05/05/making-your-python-decorators-even-better-with-functool-wraps/
-
 '''
 import argparse
 import json
